fix: 하트 누르면 슬라이드 초기화 이슈 해결
